마이페이지에 로그인 api에서 받아온 userId 넘겨줌

+++ app/src/main/java/com/example/umc_hackathon/auth/JoinResponse.kt
@@ -11,5 +11,5 @@ data class JoinResponse (
 
 data class JoinResult (
     @SerializedName("jwt") val jwt: String,
-    @SerializedName("userId") val userId: Int
+    @SerializedName("userId") val userId: Long
 )
